feat: add cw-common
