+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "contracts/cosmwasm-vm/cw-icon-light-client",
     "contracts/cosmwasm-vm/cw-xcall",
     "contracts/cosmwasm-vm/cw-ibc-core",
+    "contracts/cosmwasm-vm/cw-common",
     "libraries/rust/common"
 ]
