+++ contracts/cosmwasm-vm/cw-common/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "cw-common"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+cosmwasm-schema = "1.2.1"
+cosmwasm-std = "1.2.2"
+ibc = { version = "0.32.0", default-features = false, features = ["parity-scale-codec", "mocks-no-std", "serde", "borsh"]}
+serde = { version = "1.0.154", default-features = false, features = ["derive"] }
+cw-storage-plus = "1.0.1"
+
+[dev-dependencies]
+getrandom = {version = "0.2", default-features = false, features = ["custom"]}
