+++ contracts/cosmwasm-vm/cw-common/src/errors.rs
@@ -1,0 +1,12 @@
+use crate::types::ClientType;
+use ibc::core::ics24_host::error::ValidationError;
+
+#[derive(Debug)]
+pub enum CwErrors {
+    FailedToCreateClientId {
+        client_type: ClientType,
+        counter: u64,
+        validation_error: ValidationError,
+    },
+    InvalidClientId(ValidationError),
+}
