+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics02_client/mod.rs
@@ -6,15 +6,8 @@ pub mod handler;
 pub mod types;
 pub use super::*;
 
-use crate::types::{
-    CreateClientResponse, MisbehaviourResponse, UpdateClientResponse, UpgradeClientResponse,
-};
 use crate::{
-    context::CwIbcCoreContext,
-    msg::*,
-    traits::IbcClient,
-    types::{ClientId, ClientType},
-    ContractError, IbcClientId, IbcClientType,
+    context::CwIbcCoreContext, msg::*, traits::IbcClient, ContractError, IbcClientId, IbcClientType,
 };
 use common::icon::icon::lightclient::v1::{
     ClientState as RawClientState, ConsensusState as RawConsensusState,
@@ -23,6 +16,10 @@ use cosmwasm_std::{
     from_binary, to_binary, to_vec, Addr, CosmosMsg, DepsMut, Event, MessageInfo, Reply, Response,
     Storage, SubMsg,
 };
+use cw_common::client_response::{
+    CreateClientResponse, MisbehaviourResponse, UpdateClientResponse, UpgradeClientResponse,
+};
+use cw_common::types::{ClientId, ClientType};
 use events::{create_client_event, update_client_event, upgrade_client_event};
 use ibc::core::ics02_client::msgs::misbehaviour::MsgSubmitMisbehaviour;
 use ibc::core::{
