+++ contracts/cosmwasm-vm/cw-ibc-core/src/types.rs
@@ -11,142 +11,6 @@ use ibc::{
 use super::*;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct ClientId(IbcClientId);
-
-impl Default for ClientId {
-    fn default() -> Self {
-        Self(IbcClientId::default())
-    }
-}
-impl From<IbcClientId> for ClientId {
-    fn from(value: IbcClientId) -> Self {
-        Self(value)
-    }
-}
-impl FromStr for ClientId {
-    type Err = ibc::core::ics24_host::error::ValidationError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let r = IbcClientId::from_str(s)?;
-        Ok(Self(r))
-    }
-}
-impl ClientId {
-    /// Get this identifier as a borrowed `&str`
-    pub fn as_str(&self) -> &str {
-        self.0.as_str()
-    }
-
-    /// Get this identifier as a borrowed byte slice
-    pub fn as_bytes(&self) -> &[u8] {
-        self.0.as_bytes()
-    }
-
-    pub fn new(client_type: ClientType, counter: u64) -> Result<Self, ContractError> {
-        match IbcClientId::new(client_type.client_type(), counter) {
-            Ok(result) => Ok(Self(result)),
-            Err(error) => Err(ContractError::IbcClientError {
-                error: ClientError::ClientIdentifierConstructor {
-                    client_type: client_type.client_type(),
-                    counter,
-                    validation_error: error,
-                },
-            }),
-        }
-    }
-
-    pub fn ibc_client_id(&self) -> &IbcClientId {
-        &self.0
-    }
-
-    pub fn from(client_id: IbcClientId) -> Self {
-        Self(client_id)
-    }
-}
-
-impl<'a> PrimaryKey<'a> for ClientId {
-    type Prefix = ();
-
-    type SubPrefix = ();
-
-    type Suffix = Self;
-
-    type SuperSuffix = Self;
-
-    fn key(&self) -> Vec<Key> {
-        vec![Key::Ref(self.0.as_bytes())]
-    }
-}
-
-impl KeyDeserialize for ClientId {
-    type Output = ClientId;
-
-    fn from_vec(value: Vec<u8>) -> cosmwasm_std::StdResult<Self::Output> {
-        let result = String::from_utf8(value)
-            .map_err(StdError::invalid_utf8)
-            .unwrap();
-        let client_id = IbcClientId::from_str(&result).unwrap();
-        Ok(ClientId(client_id))
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct ClientType(IbcClientType);
-
-impl ClientType {
-    pub fn new(cleint_type: String) -> ClientType {
-        ClientType(IbcClientType::new(cleint_type))
-    }
-    pub fn client_type(&self) -> IbcClientType {
-        self.0.clone()
-    }
-
-    /// Get this identifier as a borrowed `&str`
-    pub fn as_str(&self) -> &str {
-        self.0.as_str()
-    }
-}
-
-impl From<IbcClientType> for ClientType {
-    fn from(value: IbcClientType) -> Self {
-        Self(value)
-    }
-}
-
-impl From<ClientId> for ClientType {
-    fn from(value: ClientId) -> Self {
-        let data: Vec<&str> = value.as_str().split("-").collect();
-        ClientType::new(data[0].to_string())
-    }
-}
-
-impl<'a> PrimaryKey<'a> for ClientType {
-    type Prefix = ();
-
-    type SubPrefix = ();
-
-    type Suffix = Self;
-
-    type SuperSuffix = Self;
-
-    fn key(&self) -> Vec<Key> {
-        vec![Key::Ref(self.0.as_str().as_bytes())]
-    }
-}
-
-impl KeyDeserialize for ClientType {
-    type Output = ClientType;
-
-    fn from_vec(value: Vec<u8>) -> cosmwasm_std::StdResult<Self::Output> {
-        let result = String::from_utf8(value)
-            .map_err(StdError::invalid_utf8)
-            .unwrap();
-        let client_type = IbcClientType::new(result);
-        Ok(ClientType(client_type))
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
 pub struct ConnectionId(IbcConnectionId);
 
 impl FromStr for ConnectionId {
@@ -416,144 +280,6 @@ impl From<IbcPortId> for PortId {
 impl From<IbcModuleId> for ModuleId {
     fn from(module: IbcModuleId) -> Self {
         ModuleId(module.to_string())
-    }
-}
-
-#[cw_serde]
-pub struct CreateClientResponse {
-    client_type: String,
-    height: String,
-    client_state_commitment: Vec<u8>,
-    consensus_state_commitment: Vec<u8>,
-}
-
-impl CreateClientResponse {
-    pub fn new(
-        client_type: String,
-        height: String,
-        client_state_commitment: Vec<u8>,
-        consensus_state_commitment: Vec<u8>,
-    ) -> Self {
-        Self {
-            client_type,
-            height,
-            client_state_commitment,
-            consensus_state_commitment,
-        }
-    }
-    pub fn client_type(&self) -> ClientType {
-        ClientType::new(self.client_type.to_owned())
-    }
-
-    pub fn height(&self) -> Height {
-        Height::from_str(&self.height).unwrap()
-    }
-
-    pub fn client_state_commitment(&self) -> &[u8] {
-        &self.client_state_commitment
-    }
-    pub fn consensus_state_commitment(&self) -> &[u8] {
-        &self.consensus_state_commitment
-    }
-}
-
-#[cw_serde]
-pub struct UpdateClientResponse {
-    height: String,
-    client_id: String,
-    client_state_commitment: Vec<u8>,
-    consensus_state_commitment: Vec<u8>,
-}
-
-impl UpdateClientResponse {
-    pub fn new(
-        height: String,
-        client_id: String,
-        client_state_commitment: Vec<u8>,
-        consensus_state_commitment: Vec<u8>,
-    ) -> Self {
-        Self {
-            height,
-            client_id,
-            client_state_commitment,
-            consensus_state_commitment,
-        }
-    }
-    pub fn height(&self) -> Height {
-        Height::from_str(&self.height).unwrap()
-    }
-
-    pub fn client_state_commitment(&self) -> &[u8] {
-        &self.client_state_commitment
-    }
-    pub fn consensus_state_commitment(&self) -> &[u8] {
-        &self.consensus_state_commitment
-    }
-    pub fn client_id(&self) -> Result<ClientId, ContractError> {
-        ClientId::from_str(&self.client_id).map_err(|error| ContractError::IbcDecodeError {
-            error: error.to_string(),
-        })
-    }
-}
-#[cw_serde]
-pub struct UpgradeClientResponse {
-    client_id: String,
-    height: String,
-    client_state_commitment: Vec<u8>,
-    consesnus_state_commitment: Vec<u8>,
-}
-
-impl UpgradeClientResponse {
-    pub fn new(
-        client_state_commitment: Vec<u8>,
-        consesnus_state_commitment: Vec<u8>,
-        client_id: String,
-        height: String,
-    ) -> Self {
-        {
-            Self {
-                height,
-                client_id,
-                client_state_commitment,
-                consesnus_state_commitment,
-            }
-        }
-    }
-
-    pub fn client_id(&self) -> Result<ClientId, ContractError> {
-        ClientId::from_str(&self.client_id).map_err(|error| ContractError::IbcClientError {
-            error: ClientError::InvalidClientIdentifier(error),
-        })
-    }
-
-    pub fn client_state_commitment(&self) -> &[u8] {
-        &self.client_state_commitment
-    }
-    pub fn consesnus_state_commitment(&self) -> &[u8] {
-        &self.consesnus_state_commitment
-    }
-    pub fn height(&self) -> Height {
-        Height::from_str(&self.height).unwrap()
-    }
-}
-
-#[cw_serde]
-pub struct MisbehaviourResponse {
-    client_id: String,
-    pub client_state_commitment: Vec<u8>,
-}
-
-impl MisbehaviourResponse {
-    pub fn new(client_id: String, client_state_commitment: Vec<u8>) -> Self {
-        Self {
-            client_id,
-            client_state_commitment,
-        }
-    }
-    pub fn client_id(&self) -> Result<ClientId, ContractError> {
-        ClientId::from_str(&self.client_id).map_err(|error| ContractError::IbcClientError {
-            error: ClientError::InvalidClientIdentifier(error),
-        })
     }
 }
