+++ contracts/cosmwasm-vm/cw-ibc-core/tests/channel/mod.rs
@@ -4,9 +4,9 @@ use crate::channel::test_timeout::get_dummy_raw_msg_timeout_on_close;
 use cosmwasm_std::from_binary;
 use cosmwasm_std::testing::mock_env;
 use cosmwasm_std::{Binary, IbcEndpoint, IbcPacket, IbcTimeout, IbcTimeoutBlock};
+use cw_common::types::ClientId;
 use cw_ibc_core::ics04_channel::{PacketData, PacketDataResponse};
 use cw_ibc_core::ics24_host::commitment::compute_packet_commitment;
-use cw_ibc_core::types::ClientId;
 use cw_ibc_core::IbcChannelId;
 use ibc::core::ics03_connection::connection::Counterparty as ConnectionCounterparty;
 use ibc::core::ics03_connection::connection::State as ConnectionState;
