+++ contracts/cosmwasm-vm/cw-ibc-core/tests/test_channel.rs
@@ -1,13 +1,13 @@
 use std::{str::FromStr, time::Duration};
 
 use cosmwasm_std::{to_binary, to_vec, Addr, Event, Reply, SubMsgResponse, SubMsgResult};
+use cw_common::types::ClientType;
 use cw_ibc_core::ics02_client::types::{ClientState, ConsensusState};
 use cw_ibc_core::ics04_channel::open_init::{
     create_channel_submesssage, on_chan_open_init_submessage,
 };
 use cw_ibc_core::ics04_channel::open_try::on_chan_open_try_submessage;
 use cw_ibc_core::ics04_channel::{EXECUTE_ON_CHANNEL_OPEN_INIT, EXECUTE_ON_CHANNEL_OPEN_TRY};
-use cw_ibc_core::types::ClientType;
 use cw_ibc_core::{
     context::CwIbcCoreContext,
     ics04_channel::{
