+++ contracts/cosmwasm-vm/cw-icon-light-client/Cargo.toml
@@ -43,7 +43,7 @@ prost-types = { version = "0.11.8", default-features = false }
 hex="*"
 bytes="*"
 common={path="../../../libraries/rust/common"}
-cw-ibc-core={path="../cw-ibc-core",features=["library"]}
+cw-common = { path="../cw-common" }
 
 
 [dev-dependencies]
