+++ contracts/cosmwasm-vm/cw-icon-light-client/src/bin/schema.rs
@@ -1,6 +1,6 @@
 use cosmwasm_schema::write_api;
 
-use common::client_msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use cw_common::client_msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
 
 fn main() {
     write_api! {
