+++ contracts/cosmwasm-vm/cw-icon-light-client/src/helpers.rs
@@ -1,8 +1,8 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use common::client_msg::ExecuteMsg;
 use cosmwasm_std::{to_binary, Addr, CosmosMsg, StdResult, WasmMsg};
+use cw_common::client_msg::ExecuteMsg;
 
 /// CwTemplateContract is a wrapper around Addr that provides a lot of helpers
 /// for working with this.
