+++ libraries/rust/common/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod btp_header;
-pub mod client_msg;
 pub mod icon;
 pub mod rlp;
 pub mod types;
